+++ CHANGELOG.md
@@ -32,6 +32,7 @@ The types of changes are:
 * Add Okta support to the `/generate` endpoint [#842](https://github.com/ethyca/fides/pull/842)
 * Add db support to `/generate` endpoint [849](https://github.com/ethyca/fides/pull/849)
 * Added OpenAPI TypeScript client generation for the UI app. See the [README](/clients/admin-ui/src/types/api/README.md) for more details.
+* Standardized API error parsing under `~/types/errors`
 
 ### Changed